nvim: Significant overhaul of configs
