+++ nvim/lua/plugins/lualine.lua
@@ -1,0 +1,13 @@
+return {
+  {
+    'nvim-lualine/lualine.nvim',
+    opts = {
+      options = {
+        theme = 'material',
+        component_separators = { left = '|', right = '|' },
+        section_separators = { left = '', right = '' },
+      }
+    },
+  }
+}
+
